requirements.txt
exchangelib==5.1.0
python-dotenv==1.0.0
tqdm==4.66.1